Refactor user table to separate component
